fixes some potential data loss silencing a few warnings

+++ Source/DOM classes/Core DOM/NodeList.m
@@ -19,12 +19,12 @@
     [super dealloc];
 }
 
--(Node*) item:(int) index
+-(Node*) item:(NSUInteger) index
 {
 	return [self.internalArray objectAtIndex:index];
 }
 
--(long)length
+-(NSUInteger)length
 {
 	return [self.internalArray count];
 }
